Added Admin Role Editing
In the User Index, admin can now edit the roles of users, either adding or removing the admin or editor role. Potential bug: the admin can remove their own admin role, which should not be allowed.
We could limit adding the admin role to an admin user, but that button is already gated by the user index only being accessible to an admin.