Manage Users Functionality Complete
Everything about managing users has been completed assign from 1 slight behavior bug: when an admin creates a user, they are then signed out of their account and then signed in as the user. It might be better to have a confirmation page declaring the user has been created, and then the details of the user that was created, including the temporary password, and a reminder to remind the person to change their temporary password.