7/8 Registration Tests Pass
The only test that does not pass is Registration>Create>Post. It throws a null reference exception when attempting to assign to tempdata. Really not sure why, probably some intricacy of .net core mvc. I'm going to try and bypass it by passing the registration being created directly to the create confirmation controller, instead of passing strings to TempData. Hold my beer.