Organization Creation View Overhaul
Working towards tag inclusion on organization creation.
Step 1: Create form-fields that can be dynamically added to the page.
Step 2: Include a list of tags in the view
Step 3: Set the dynamic form fields to be selectors for tags.
Step 4: On for submission, save the contents of each dropdown to the list of tags.

After reviewing this process, it might be easier to have one dropdown rather than multiple, and then a button that dynamically pulls from the box and adds to the list of tags. It'd be easier to code, too.