Bugfix: CAP-307, CAP-292
CAP-307: Creating an article or FAQ no longer fails due to not being able to duplicate a category.
CAP-292: Not completely fixed; anonymous users can now click the back breadcrumb in article details, but instead of viewing articles with the same category, it instead returns to a list of categories. More work is needed to fix CAP-292. Added a possible solution as a comment to CAP-292.