Added Contact Form Functionality
This implementation uses the DotNetEmailService class to get an email message from an anonymous user, and send it to the email address specified in the startup class.
I also added some validation type stuff using this guide: https://dotnetthoughts.net/how-to-use-bootstrap-style-validation-in-aspnet-core/
Note how it is used in the ContactUs view; an asp validation tag helper is used just like normal, but instead of 'class=text-danger', we use 'class="invalid-tooltip."'
This will actually highlight the boxes in red if they are invalid and add an x at the end, or green with check at the end if they are valid. It works alongside what we already have.

I also wen through and updated "Contact Us" links to point to the right view, and added a button in the header that links to the contact form as well.

I also added a Notification view that works like the CustomError view I made earlier, but with two differences.
1: The text is black, not red since it is not an error.
2: The 'Go Back' button is replaced with a 'Home' button.

I also made it so that emails sent through the emailer are treated as html, so it's possible for us to style emails however we'd like. Ideally, we'd have specific boilerplate html email files that we could feed to the emailer, but for now we have to put html in strings when sending an email (see AboutController - ContactUs [POST] method).