Encapsulated Form Button Tag Helpers
Now it will be very easy to implement a form button tag helper, which makes it very easy to implement form buttons. See "FormSendButtonTagHelper" to see how I easily implemented the tag-helper, and look at the ContactUs view form submit button to see how easy it was to implement the submit button in cshtml.