getting model styling uniform ready
It looks like I did a lot more than I actually did. One of the biggest file changes you see is the result of changing the name of the Training Program controller from TrainingProgramsController to TrainingProgramController. It's subtle, but that extra s was not in line with the rest of our controllers.

The real meat of what I did was as follows:
- Finish encapsulating the tag helpers
- Create a CRUD layout that can be used across our models.
     - I chose to use payment options as a guinea pig.